pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lvrgd-common"
version = "0.1.19"
description = "Common services and utilities for LVRGD projects"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "LVRGD Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "boto3>=1.35.0",
    "loguru>=0.7.3",
    "minio>=7.2.7",
    "motor>=3.3.0",
    "mypy-boto3-dynamodb>=1.40.56",
    "pydantic>=2.11.9",
    "pymongo>=4.15.1",
    "pynamodb>=6.1.0",
    "redis>=5.0.0",
    "rich>=14.2.0",
]

[project.optional-dependencies]
dev = [
    "mongomock>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.0.0",
    "ruff>=0.13.1",
    "flake8>=7.3.0",
    "black>=25.9.0",
]

[tool.hatch.build]
include = [
    "src/lvrgd/**/*.py",
    "src/lvrgd/py.typed",
]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.uv]
package = true

[dependency-groups]
dev = [
    "boto3-stubs[dynamodb]>=1.35.0",
    "mongomock>=4.3.0",
    "pymongo>=4.15.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 100
extend-exclude = [".github", "playground"]
lint.select = ["ALL"]
lint.ignore = [
    "COM812",  # Conflicts with formatter
    "D203",    # Conflicts with D211
    "D213",    # Conflicts with D212
    "E501",    # Line too long (handled by formatter)
    "TC001",   # Type-checking block - overly strict, runtime imports are fine
    "TC003",   # Type-checking block - overly strict, runtime imports are fine
    "PLR0913", # Too many arguments - 5 is arbitrary, 6-7 args acceptable for clarity
    "ANN401",  # Disallow Any - sometimes needed for dict(**kwargs) patterns
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "D100",    # Don't require module docstrings in tests
    "D103",    # Don't require function docstrings in tests
    "S101",    # Allow assert statements in tests
    "S105",    # Allow hardcoded passwords in tests
    "S106",    # Allow hardcoded password arguments in tests
    "S108",    # Allow temp file usage in tests
    "SLF001",  # Allow private member access in tests
    "PLR2004", # Allow magic values in tests
    "INP001",  # Tests don't need __init__.py
    "ARG001",  # Unused function arguments in fixtures
    "ARG002",  # Unused method arguments in test methods
]
"integration-tests/**/*.py" = [
    "D100",    # Don't require module docstrings in integration tests
    "D103",    # Don't require function docstrings in integration tests
    "S101",    # Allow assert statements in integration tests
    "S105",    # Allow hardcoded passwords in integration tests
    "S106",    # Allow hardcoded password arguments in integration tests
    "S108",    # Allow temp file usage in integration tests
    "S306",    # Allow insecure mktemp in integration tests
    "SLF001",  # Allow private member access in integration tests
    "PLR2004", # Allow magic values in integration tests
    "INP001",  # Integration tests don't need __init__.py
    "ARG001",  # Unused function arguments in fixtures
    "ANN001",  # Allow missing type annotation for logger fixture
    "N999",    # Allow integration-tests module name with hyphen
    "TRY003",  # Allow string literals in exceptions for test assertions
    "TRY301",  # Allow raise in try blocks for testing error handling
    "EM101",   # Allow string literals in exceptions for test messages
    "B011",    # Allow assert False in tests
    "PT015",   # Allow assert False in exception tests
]
"**/minio_models.py" = [
    "S105",    # Example credentials in Field examples
]
"**/redis_service.py" = [
    "FBT001",  # Boolean arguments required for Redis API compatibility
    "FBT002",  # Boolean default arguments required for Redis API compatibility
    "A003",    # 'set' method name required for Redis API compatibility
    "TRY300",  # Return in try block is clearer in this context
]
"**/async_redis_service.py" = [
    "FBT001",  # Boolean arguments required for Redis API compatibility
    "FBT002",  # Boolean default arguments required for Redis API compatibility
    "A003",    # 'set' method name required for Redis API compatibility
    "TRY300",  # Return in try block is clearer in this context
]
"**/redis_models.py" = [
    "FBT003",  # Boolean positional values in Pydantic Field acceptable
]
"**/dynamodb_service.py" = [
    "PLR0911", # Too many return statements acceptable for operator mapping
    "TRY003",  # Long exception messages acceptable for error context
    "EM102",   # F-string in exceptions acceptable for error context
    "TRY300",  # Return in try block is clearer in this context
    "TRY301",  # Raise in try block is clearer in this context
    "TRY400",  # logging.error with raise is intentional for structured logging
]
"**/sort_key_condition.py" = [
    "TRY003",  # String literal in exception acceptable for validation
    "EM101",   # String literal in exception acceptable for validation
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["lvrgd"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
